+++ tools/go-agent/instrument/plugins/register.go
@@ -20,6 +20,7 @@ package plugins
 import (
 	"github.com/apache/skywalking-go/plugins/core/instrument"
 	"github.com/apache/skywalking-go/plugins/dubbo"
+	"github.com/apache/skywalking-go/plugins/fasthttp"
 	"github.com/apache/skywalking-go/plugins/gin"
 	goredisv9 "github.com/apache/skywalking-go/plugins/go-redisv9"
 	"github.com/apache/skywalking-go/plugins/go-restfulv3"
@@ -53,6 +54,7 @@ func init() {
 	registerFramework(mux.NewInstrument())
 	registerFramework(grpc.NewInstrument())
 	registerFramework(irisv12.NewInstrument())
+	registerFramework(fasthttp.NewInstrument())
 
 	// gorm related instruments
 	registerFramework(gorm_entry.NewInstrument())