+++ CHANGES.md
@@ -10,6 +10,7 @@ Release Notes.
 * Support [mux](https://github.com/gorilla/mux) HTTP server framework.
 * Support [grpc](https://github.com/grpc/grpc-go) server and client framework.
 * Support [iris](https://github.com/kataras/iris) framework.
+* Support [fasthttp](https://github.com/valyala/fasthttp) framework.
 
 #### Documentation